backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiohttp==3.9.1
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.2
python-telegram-bot==20.7
ffmpeg-python==0.2.0
pysrt==1.1.2
webvtt-py==0.4.6
ass==0.5.4
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
streamlink==6.4.2
yt-dlp==2023.11.16
m3u8==4.0.0
cryptography==41.0.7
pillow==10.1.0
numpy==1.26.2